An attempt to integerate simulus This is far from perfect, but it works (probably not "correctly", however). There is a bug where the calculated death time ends up being before the current time due to the sugar being in the negatives. The problem is mostly that the death should be scheduled before the sugar levels even get <= 0...
A temporary "bandage" is added in agent.py line 308, which shedules the death at the time when it is detecdted that the sugar level is <= 0. This means, however, that an Agent may execute another event whilst being in the negatives of sugar -- this is obviously not a correct way to simulate this. I've spent all day trying to figure this out but I couldn't figure it out, perhaps I'll be able to after a good night's rest. For now, it'll stay like this for maybe a day or so.

Overall, the major changes have occured in event.py:

EventCalendar:
- No longer keeps its own Event list.
- A few wrapper functions for simulator calls.
- Decided not to use simulator.run() because the pre and post event procedures (like PAUSE, SHOW_ANIMATION, landscape sugar update), rather essentially reimplimented the same thing, but allows for such operations.

Event:
- callback member added to hold the function that is to be executed when scheduled by simulus

Also major changes sustained with pause.py where some commands were removed because they would not be immediately compatible with simulus (I might add them back later). Agent no longer has setNextEvent, getNextEvent, and other functions that were meant to manage its own "mini calendar", rather it just schedules events wihtout much restriction.